Tokenizer: The memory for `last_start_tag_name` is now externally managed
The `last_start_tag_name` is only used when emitting an end tag token while in a state that is not `Data` nor `CDATASection`.
However, the tokenizer can only be in such a state when its state is changed externally; i.e. it can never be in a state that is not `Data` or `CDATASection` on its own.
Therefore, we can make it the parser's responsibility to set `last_start_tag_name`.
Furthermore, the Tokenizer no longer needs to set `last_start_tag_name` whenever it emits a start tag.
This means `last_start_tag_name` can point to static memory and not heap-allocated memory.